requirements.txt
discord.py
asyncio
configparser
SQLAlchemy
dislash.py
wavelink
art
aiohttp
git+git://github.com/Rapptz/discord-ext-menus
humanize
requests
ujson
psycopg2
pycountry
six
google-cloud-translate>=3.0.0
tortoise-orm
